Improved Tree Info, Reconfigured Control Scaling
Changed the Tree display to be more consistent in what information it displays. It now always displays elements in the format "[AA] NAME (YY)", where AA is the element's ID in hex, NAME is the element's name, and YY is the element's size in hex. In doing this, the display for sounds themselves now use their proper names.

Additionally, configured the anchoring rules such that, when resizing the window, the console and "Insert" controls will scale up/down, leaving the Tree view a constant size.

Lastly, disabled auto-play on multi-channel audio, as they I'm pretty sure always play garbage.