Modified LoadTreeView For Proper Name Support
The previous addition of names to the TreeView broke things a bit in that sometimes sounds would display the incorrect name and ID. This was caused by an oversight on my part: I forgot to consider that the TreeView was built from an RWSD's WAVE section, while the list of names I was pulling from was for an RWSD's DATA section.

To rectify this, I've modified the LoadTreeView function to instead use the DATA section list, allowing the names to properly sync up. Additionally, it should be noted that the bracketed numbers accompanying the sound names are still Wave IDs, same as they have always been. It's now possible, however, that you'll see duplicates as a result of multiple sounds sharing a single Wave ID. You can actually see the same in BrawlCrate if you right click a sound and open the "View File" dialog.

Note: In doing this, I've also added support for viewing WAVE entries which have been left orphaned somehow (eg. by use of lavaSawndPortingUtility), which is neat.

Thanks to SoN1c2001/Burrotis for bringing this to my attention!