Reimplemented Multi Export List, Added Tooltips
Reworked the Groups list in the Multi Sawnd Export to use a TreeView control instead of a CheckedListView. This allows single clicks toggling checkboxes to be isolated to just the actual checkbox rather than triggering on the entire entry, while maintaining the ability to double-click or press enter on an entry to toggle its check box.

Additionally, tooltips have been added to the Naming Scheme options explaining what they do.

Lastly, added a "Cancel" button to the Multi Export dialog, and enabled closing both that dialog and the Manual Naming dialog with the Escape key.