Adds a Check To Avoid Invalid InfoIndex Requests
This wasn't necessary before, but because it's now possible to replace BRWSDs with ones that have different numbers of data entries, it can sometimes be the case that imported entries don't get assigned Info Indeces. In the event that this happens, it'll be impossible to edit the sound's properties, which is indicated by a Yellow background.

Also fixes an issue where the TreeView wasn't properly being refreshed after importing a BRWSD.