Reimplemented Search Functionality
Search functionality now uses internal string dictionaries for searching, ditching the trie-based approach I attempted before. While that did work, it just wasn't as intuitive as I'd wanted, as it required that the searched text be at the beginning of the word. I think an approach like that might be useful at some point for an autocomplete to this new search bar though.

This implementation also allows for the user to change the way the search is conducted, and even allows for RegEX based searching, which is quite nifty I think.