Reimplemented Auto Mono Conversions
Conversions down from Stereo WAVs to Mono are now controlled via a setting, which can be changed through the Settings Menu, and can be set to always convert to mono, only convert when replacing a mono sound (the default), or to never convert.

Conversions are still handled by VGAudio, which I believe should be fine ultimately since the conversion should be lossless.

In the future, I think I'll see about expanding this functionality to also be able to downsample sounds automatically, but I'll need to find another library to handle that I think. I also wanna look into having the program be able to accept other more common sound formats, but that'll be a whole other undertaking since it'd require other codecs (e.g LAME) be installed and checked for.