Initial Treeview Tabs Implementation
There are now a pair of tabs, which you can move groups between by using the groups' right-click menus.

This actually kinda fakes its tabs: the panes for the respective tabs are actually set to 0 height, we don't use them. The only thing we're using it for is selecting the desired list: when the selected tab changes, we simply swap out the nodes currently stored in the TreeView.