Fixed Sawnd Drag and Drop Imports
I'd previously added a check to try and prevent imports when no BRSAR was loaded by checking whether or not the normal insert button was enabled. This was silly of me actually, because all controls get disabled while drag-dropping, so it would always return false. This new iteration checks properly whether a BRSAR is loaded and disables it if one isn't. Not sure why I didn't do this to begin with.