Tab Settings Now Only Reset on BRSAR Load
Made some changes to ensure that the current tab settings are preserved through loads of the tree view unless either a new BRSAR is loaded or the tab configuration is updated through the settings menu.

Prevents an issue where, if the user had made any changes to the tab layout without saving, those changes would be erased if the treeview was reloaded.