Added Sorting Options to Bulk File Exporter
Additionally reworked how data is handled under the hood, we now maintain an internal list of the files in consideration, and sort that as necessary. We then build the tree view according to that, and store the corresponding collection's index in that list in each node's Tag field (using that index to access Files from the list).

Also, the number of checked collections is now reported to the user.