Made It Clearer When Tab Settings Are Saved
Added a checkbox to the settings menu which denotes whether or not the current tab settings will be saved along with all other settings. If not checked, they will be left alone, whereas if it's checked it'll overwrite the current settings. Note that the checkbox will check itself automatically if you import any tab settings!