Potential WAV Import Pop Fix
What we're doing here is just setting the first two bytes of every block to zero. It appears that in certain circumstances, the encoding function sets those two bytes high even when the original wav is silent at that point, which seems to be what's causing the pop.

Admittedly, I don't understand *why* the encoding function is doing this, but for now at least, this looks like a fix.