Separated Preproc Audio Trim Into Its Own Option
This makes it possible to trim and pad in the same transaction, which makes it possible to do things like silence the beginning of a sound by trimming it off, then padding the same length back onto the front.

Additionally, added a field which reports the length of the audio post-processing.

Lastly, fixed an issue where the Padding Length counters could still use negative values.