Reworked "Allow Clipping" into "Normalize Volume"
This checkbox didn't previously do anything, and I wasn't sure exactly how to accomplish what I wanted with it. I've since learned that what I actually wanted was a quick way to maximize the volume of a sound without any clipping, and that SoX's "Normalize" function is exactly what I wanted, actually!

So, that checkbox has been reworked to serve the purpose of normalizing audio volume. It disables the Volume control itself, and applies that amplification as necessary.