WAV Preproc Form Now Respects Settings Menu
The WAV Preprocessing form is now properly named, and also respects the options in the Settings Menu regarding pre-processing audio imports!

Additionally, the singular ImportWAV function in sawndz.cs has been slightly reworked to no longer use the old doInsertWithRespectToChannelCount function! The form is now set up in such a way that initializing the form with an audio file is enough to perform all necessary audio processing, even without opening it, which I hope to use  in order facilitate the headless version of this functionality I'll need for to tie the Multi-WAV import functionality into this whole process.

Lastly, when loading the form itself, the WAV targeted for replacement will be targeted by default.