Second Attempt at WAV Import Pop Fix
There was a minor issue previously where some imported sounds would have a slight popping sound added to their fronts. Frankly, I'm still not perfectly clear on why this was happening, but one way that I found to fix it was just to silence the first sample of the audio itself. However, as I found later, this was causing issues in sounds which looped for some reason, essentially shifting the popping issue off of unlooped sounds and onto looped ones. I've poked at this for a good little while and still don't have a real explanation for *why* this issue occurs, especially so in cases where the zeroed out first sample is nowhere near the loop point. What I have found though is that just isolating the fix to non-looped sounds seems to work well enough that I'm comfortable deploying it. I do still intend to look more closely at this at a later date, now just isn't the time lol.